Create a ui module. Seperated the ui codes.
